refactor: update docker files and usecase apps to work with refactor
