+++ usecase-counter/lib.rs
@@ -1,7 +1,4 @@
-pub mod validator;
-
-pub mod node;
-pub mod router;
 pub mod types;
+pub mod validator;
 
 pub use validator::CounterValidator;
