+++ usecase-counter/node/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "commonware-avs-node-app"
+version = "0.1.0"
+edition = "2024"
+
+[[bin]]
+name = "commonware-avs-node-app"
+path = "src/app.rs"
+
+[dependencies]
+commonware-usecase-counter = { path = ".." }
+commonware-avs-node = { path = "../../node" }
+commonware-avs-router = { path = "../../router" }
+commonware-eigenlayer = { workspace = true }
+commonware-cryptography = { workspace = true }
+commonware-runtime = { workspace = true }
+commonware-p2p = { workspace = true }
+commonware-utils = { workspace = true }
+clap = { workspace = true }
+ark-bn254 = { workspace = true }
+bn254 = { workspace = true }
+tracing = { workspace = true }
+dotenv = { workspace = true }
+eigen-logging = { workspace = true }
+governor = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+commonware-avs-core = { path = "../../core" }
