+++ usecase-counter/router/src/creator.rs
@@ -3,8 +3,8 @@ use async_trait::async_trait;
 use std::sync::Arc;
 use tracing::error;
 
-use crate::router::provider::CounterProvider;
-use crate::types::CounterTaskData;
+use crate::provider::CounterProvider;
+use commonware_usecase_counter::types::CounterTaskData;
 use commonware_avs_router::creator::queue::{CreatorConfig, SimpleTaskQueue, TaskQueue};
 use commonware_avs_router::ingress::types::TaskRequest;
