+++ usecase-counter/router/src/factories.rs
@@ -1,4 +1,4 @@
-use crate::router::{
+use crate::{
     creator::{CounterCreator, CounterCreatorType, ListeningCounterCreator},
     executor::CounterHandler,
     provider::CounterProvider,
