+++ usecase-counter/types.rs
