+++ usecase-counter/validator.rs
